+++ src/test/java/com/iancheng/ecommerce/controller/ProductControllerTest.java
@@ -13,6 +13,8 @@ import org.springframework.test.web.servlet.RequestBuilder;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
+
 import static org.hamcrest.Matchers.*;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
@@ -63,7 +65,7 @@ public class ProductControllerTest {
         productRequest.setProductName("test food product");
         productRequest.setCategory(ProductCategory.FOOD);
         productRequest.setImageUrl("http://test.com");
-        productRequest.setPrice(100);
+        productRequest.setPrice(BigDecimal.valueOf(100));
         productRequest.setStock(2);
 
         String json = objectMapper.writeValueAsString(productRequest);
@@ -110,7 +112,7 @@ public class ProductControllerTest {
         productRequest.setProductName("test food product");
         productRequest.setCategory(ProductCategory.FOOD);
         productRequest.setImageUrl("http://test.com");
-        productRequest.setPrice(100);
+        productRequest.setPrice(BigDecimal.valueOf(100));
         productRequest.setStock(2);
 
         String json = objectMapper.writeValueAsString(productRequest);
@@ -157,7 +159,7 @@ public class ProductControllerTest {
         productRequest.setProductName("test food product");
         productRequest.setCategory(ProductCategory.FOOD);
         productRequest.setImageUrl("http://test.com");
-        productRequest.setPrice(100);
+        productRequest.setPrice(BigDecimal.valueOf(100));
         productRequest.setStock(2);
 
         String json = objectMapper.writeValueAsString(productRequest);
